// typescript/trading-economics-dashboard/types/financial-data.ts
export interface FinancialData {
  Symbol: string;
  Date: string;
  last: number;
  value1: number;
  date1: string;
  value2: number;
  date2: string;
  value3: number;
  date3: string;
  value4: number;
  date4: string;
  unit: string;
  currency: string;
  frequency: string;
  stock: string;
  financialSymbol: string;
}